+++ src/Areas/Storage/Options/DataLake/FileSystem/ListPathsOptions.cs
@@ -1,0 +1,6 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+namespace AzureMcp.Areas.Storage.Options.DataLake.FileSystem;
+
+public class ListPathsOptions : BaseFileSystemOptions;