+++ src/Areas/Storage/Options/StorageOptionDefinitions.cs
@@ -8,6 +8,7 @@ public static class StorageOptionDefinitions
     public const string AccountName = "account-name";
     public const string ContainerName = "container-name";
     public const string TableName = "table-name";
+    public const string FileSystemName = "file-system-name";
 
     public static readonly Option<string> Account = new(
         $"--{AccountName}",
@@ -28,6 +29,14 @@ public static class StorageOptionDefinitions
     public static readonly Option<string> Table = new(
         $"--{TableName}",
         "The name of the table to access within the storage account."
+    )
+    {
+        IsRequired = true
+    };
+
+    public static readonly Option<string> FileSystem = new(
+        $"--{FileSystemName}",
+        "The name of the Data Lake file system to access within the storage account."
     )
     {
         IsRequired = true